refactor: move RegionalAnalyticsView to its own file
Closes #139
